Add files via upload
Basic logic configuration for having a multi-element page layout, wherein one element is used for User input, and that input is logged with prefix of timestamp formatted in 24-hour clock.  The user input will scroll if overflows, and when overflows, the element scrolling is always positioned at the BOTTOM of the scrolling (i.e. end of last message element aligned with bottom of scrolling element).